# test: use t.Parallel to speed up pruning tests