--- a/store/pruning/manager_test.go
+++ b/store/pruning/manager_test.go
@@ -79,12 +79,14 @@ func TestStrategies(t *testing.T) {
 		},
 	}
 
-	manager := pruning.NewManager(db.NewMemDB(), log.NewNopLogger())
-
-	require.NotNil(t, manager)
-
 	for name, tc := range testcases {
+		tc := tc // Local copy to avoid shadowing.
 		t.Run(name, func(t *testing.T) {
+			t.Parallel()
+
+			manager := pruning.NewManager(db.NewMemDB(), log.NewNopLogger())
+			require.NotNil(t, manager)
+
 			curStrategy := tc.strategy
 			manager.SetSnapshotInterval(tc.snapshotInterval)