On my machine, with the race detector enabled, using t.Parallel takes the runtime for the store/pruning package from about 12s to about 2.2s. With the race detector disabled, it drops the time from about 2.6s to about 0.5s.

## Description

The `store/pruning` package includes a table-driven test, where each test case takes a noticeable time to run. However, the test cases should be independent of each other, so using `t.Parallel` makes a measurable speedup in running this package's tests.

---

### Author Checklist


I have...

- [x] included the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
- [x] targeted the correct branch (see [PR Targeting](https://github.com/cosmos/cosmos-sdk/blob/main/CONTRIBUTING.md#pr-targeting))
- [ ] ~~provided a link to the relevant issue or specification~~
- [ ] ~~reviewed "Files changed" and left comments if necessary~~
- [ ] confirmed all CI checks have passed

### Reviewers Checklist

I have...

- [ ] confirmed the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
- [ ] confirmed all author checklist items have been addressed
- [ ] confirmed that this PR does not change production code